Screenshots of errors
These are the 3 screenshots of when each command is ran. Cant seem to figure out how to make the call return the data im asking for in the functions. When each command is ran the error that returns is "this is loaded". This leads me to believe that I am just missing some links in the functions because its not throwing a total "not working" error. I feel I am getting closer to fixing it.